[core] Add pull-based visualisation service
* Add a shared visualisation backend that stores post-master PCM history and serves pull-based PCM and spectrum windows
* Sessions register requirements with the backend, keep their own channel selection, and unregister automatically when released
* Add a PFFFT wrapper for spectrum generation, and vendor PFFFT into build

+++ include/core/engine/enginedefs.h
@@ -117,7 +117,7 @@ Q_DECLARE_FLAGS(RGProcessing, RGProcess)
 Q_FLAG_NS(RGProcessing)
 
 //! Analysis frame types produced by the analysis bus.
-enum AnalysisDataType : uint8_t
+enum AnalysisDataType : uint32_t
 {
     NoAnalysisData = 0,
     LevelFrameData = 1 << 0,
